little fixes

+++ frontend/src/person/PersonMainData.js
@@ -5,7 +5,7 @@ class PersonMainData extends React.Component {
     format_date() {
         const date_string = this.props.person.birth_date;
         const date = new Date(Date.parse(date_string))
-        const date_format_options = { year: 'numeric', month: 'long', day: 'numeric' };
+        const date_format_options = { year: 'numeric', month: 'numeric', day: 'numeric' };
         return date.toLocaleDateString(undefined, date_format_options)
     }
